mergertexture: use old blocks as much as possible

+++ lib/heightmap/heightmap/render/mergertexture0.frag
@@ -1,4 +1,4 @@
-uniform highp sampler2D qt_Texture0;
+uniform mediump sampler2D qt_Texture0;
 varying mediump vec2 qt_TexCoord0;
 
 void main(void)
